+++ tests/functionnal/Config.php
@@ -38,7 +38,6 @@ namespace tests\units;
 use DbTestCase;
 use Glpi\Plugin\Hooks;
 use Log;
-use PHPMailer\PHPMailer\PHPMailer;
 use Session;
 
 /* Test for inc/config.class.php */
@@ -269,11 +268,11 @@ class Config extends DbTestCase
         $this->boolean(\Config::getLibraryDir(''))->isFalse();
         $this->boolean(\Config::getLibraryDir('abcde'))->isFalse();
 
-        $expected = realpath(__DIR__ . '/../../vendor/phpmailer/phpmailer/src');
+        $expected = realpath(__DIR__ . '/../../vendor/symfony/console');
         if (is_dir($expected)) { // skip when system library is used
-            $this->string(\Config::getLibraryDir('PHPMailer\PHPMailer\PHPMailer'))->isIdenticalTo($expected);
+            $this->string(\Config::getLibraryDir('Symfony\Component\Console\Application'))->isIdenticalTo($expected);
 
-            $mailer = new PHPMailer();
+            $mailer = new \Symfony\Component\Console\Application();
             $this->string(\Config::getLibraryDir($mailer))->isIdenticalTo($expected);
         }